Close #49
