+++ magplots/__init__.py
@@ -1,0 +1,8 @@
+"""init.py"""
+
+from importlib import metadata
+
+from magplots.magFunctions import *  # noqa F401
+
+# Set version
+__version__ = metadata.version('magplots')
