+++ notebooks/stations.csv
@@ -1,0 +1,18 @@
+IAGA,GEOLON,GEOLAT,AACGMLON,AACGMLAT,STATION-NAME,OPERATOR-NUM,OPERATOR,OP2,OP3
+THL,290.77,77.47,29.24,84.72,Qaanag,2,DTUSpace,INTERMAGNET,
+SVS,294.9,76.02,32.87,83,Savissivik,1,DTUSpace,,
+KUV,302.82,74.57,41.92,80.69,Kullorsuaq,1,DTUSpace,,
+UPN,303.85,72.78,40.2,78.93,Upernavik,1,DTUSpace,,
+UMQ,307.87,70.68,42.58,76.38,Uummannaq,1,DTUSpace,,
+GDH,306.47,69.25,39.39,75.25,Godhavn,2,DTUSpace,INTERMAGNET,
+ATU,306.43,67.93,38.19,73.99,Attu,1,DTUSpace,,
+STF,309.28,67.02,40.87,72.64,Kangerlussuaq,1,DTUSpace,,
+SKT,307.1,65.42,37.22,71.43,Maniitsoq,1,DTUSpace,,
+GHB,308.27,64.17,37.85,69.98,Nuuk,1,DTUSpace,,
+FHB,310.32,62,39.05,67.41,Paamiut,1,DTUSpace,,
+PG1,77.2,-85.5,33.68,-76.55,Antartica,2,THEMIS,MIST,
+PG2,57.96,-84.42,39.07,-75.24,Antartica,2,THEMIS,MIST,
+PG3,37.63,-84.81,36.76,-73.53,Antartica,2,THEMIS,MIST,
+PG4,12.25,-83.34,36.39,-70.83,Antartica,2,THEMIS,MIST,
+PG0,88.68,-83.67,38.23,-78.32,PG0,2,THEMIS,MIST,
+PG5,5.71,-81.96,37.17,-69.46,PG5,2,THEMIS,MIST,
