--- output/selected/selected_events.csv
@@ -1,7 +1,0 @@
-Max,Original Jet Number,Ref Spacecraft,Start,End,SM_LAT,SM_LON,SM_X,SM_Y,SM_Z,GEO_X_1,GEO_Y_1,GEO_Z_1,DIST_FROM_P93_BOW_SHOCK,DIST_FROM_MAGNETOPAUSE,DIST_FROM_T95_NS,L_VALUE,INVAR_LAT,MAGNETIC_STRENGTH,Dynamic Pressure (nPa),Electron density,Vx,Vy,Vz,Jet Number,Count
-2013-01-03 12:39:03,11,A,2013-01-03 12:38:52,2013-01-03 12:39:57,-5.78113,21.5338,10.3168,4.07093,-1.12289,10.8749,2.02144,-1.38595,-3.7506,0.178121,-1e+31,11.2619,72.6634,-1e+31,1.338730478558712,52.16679759408922,61.11369377994187,85.11691550800647,140.54807316895244,11,0
-2014-02-19 03:47:07,200,A,2014-02-19 03:46:41,2014-02-19 03:50:20,-15.4036,18.195,10.4916,3.44846,-3.04272,-9.23353,6.61546,-1.48403,-3.22347,0.507063,-1e+31,12.3249,73.4504,-1e+31,2.7189102759466532,51.9522332204102,-132.70477154109457,53.69593801767556,11.362804473045749,200,0
-2015-03-27 20:07:06,247,A,2015-03-27 20:06:39,2015-03-27 20:07:16,20.5758,23.3562,9.11086,3.93435,3.72543,-1.24167,-10.3115,2.1208,-4.35729,-0.377101,-1e+31,12.0941,73.2887,54.9513,1.5437180892042113,27.845994667032343,-230.99375345400819,114.142168171997,-2.4319949878710574,247,0
-2015-04-05 19:14:46,261,A,2015-04-05 19:14:39,2015-04-05 19:15:09,20.9746,15.3663,9.29268,2.55375,3.69447,-0.480458,-10.097,2.08439,-4.27286,-0.623951,-1e+31,11.8379,73.1034,59.5534,3.8598865013470856,31.369855545667154,-199.56062889299653,321.8777125124335,62.68103161597538,261,0
-2015-04-11 23:25:08,267,A,2015-04-11 23:24:34,2015-04-11 23:25:35,15.2148,22.3274,10.754,4.41655,3.16183,-9.9365,-6.25711,2.6957,-2.86721,1.08031,-1e+31,12.9391,73.859,-1e+31,1.584974064316785,1.4118809105903738,5.899067450904097,-70.66921127813718,16.437433584590917,267,0
-2016-05-20 08:34:10,799,A,2016-05-20 08:32:44,2016-05-20 08:34:32,-4.34393,23.3817,10.417,4.50386,-0.862081,2.68994,11.0312,0.785607,-3.7705,0.372195,-1e+31,11.4473,72.8088,-1e+31,2.64519426370169,23.112865043494125,-149.00199558956805,198.96844723822878,-33.48992066925439,799,0
