+++ pyproject.toml
@@ -26,5 +26,12 @@ classifiers = [
 Homepage = "https://github.com/KCollins/magplots"
 "Bug Tracker" = "https://github.com/KCollins/magplots/issues"
 Changelog = "https://github.com/KCollins/magplots/releases"
+source = "https://github.com/KCollins/magplots"
+download = "https://github.com/KCollins/magplots/releases"
 
-[tool.setuptools_scm]
+[tool.setuptools]
+packages = ["magplots"]
+
+[tool.pylint.'MESSAGES CONTROL']
+max-line-length = 120
+disable = "R0912,R0913,R0914,R0915,R0917,C0103,W0622"
